refactor(otlp-exporter): move code that will be removed to legacy directory
