+++ experimental/packages/opentelemetry-exporter-metrics-otlp-proto/src/index.ts
@@ -14,4 +14,5 @@
  * limitations under the License.
  */
 
-export * from './OTLPMetricExporter';
+// Legacy exporter kept for compatibility, scheduled for removal in 2.0
+export { OTLPMetricExporter } from './legacy/OTLPMetricExporter';
