+++ experimental/packages/otlp-exporter-base/src/index.ts
@@ -13,17 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './platform';
-export { OTLPExporterBase } from './OTLPExporterBase';
-export {
-  OTLPExporterError,
-  OTLPExporterConfigBase,
-  ExportServiceError,
-} from './types';
-export {
-  parseHeaders,
-  appendResourcePathToUrl,
-  appendRootPathToUrlIfNeeded,
-  configureExporterTimeout,
-  invalidTimeout,
-} from './util';
+
+// Legacy exports scheduled for removal.
+export * from './legacy';
