+++ experimental/packages/otlp-exporter-base/src/legacy/OTLPExporterBase.ts
