+++ experimental/packages/otlp-exporter-base/src/legacy/platform/browser/util.ts
