+++ experimental/packages/otlp-exporter-base/src/legacy/platform/index.ts
