+++ experimental/packages/otlp-exporter-base/src/legacy/platform/node/util.ts
