+++ experimental/packages/otlp-exporter-base/src/legacy/types.ts
