+++ experimental/packages/otlp-exporter-base/src/legacy/util.ts
