+++ experimental/packages/otlp-exporter-base/test/common/legacy/util.test.ts
@@ -22,7 +22,7 @@ import {
   appendResourcePathToUrl,
   appendRootPathToUrlIfNeeded,
   parseRetryAfterToMills,
-} from '../../src/util';
+} from '../../../src/legacy/util';
 
 describe('utils', () => {
   afterEach(() => {
