+++ experimental/packages/otlp-proto-exporter-base/src/legacy/platform/browser/index.ts
