+++ experimental/packages/otlp-proto-exporter-base/src/legacy/platform/index.ts
