+++ experimental/packages/otlp-proto-exporter-base/src/legacy/platform/node/util.ts
