+++ experimental/packages/otlp-proto-exporter-base/src/legacy/platform/types.ts
